requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
